Personalization edits
Fix #3

Update blog with relevant personalization links. The strangest piece of
the puzzle is figuring out how to add an avatar. The `hugo-cactus-theme`
requires the `avatar.png` file to be added to its static images, which
requires me to make an update and commit within
`themes/hugo-cactus-theme`. Which is a little weird, but not the end of
the world.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>