Update travis yml to resolve deploy issues
Instead of deploying through the `deploy` step in `travis`, instead
deploy during the `script` portion, but only on the `master` branch.
This specification means we will only build when we have already merged
the pull request. And I suspect it will also resolve the issues with not
being in the proper directory or having the environment variables which
seemed to be the issue before.

Additionally, we no longer store our theme as a submodule, as it makes
deploy with travis act weird. We'll examine how to fix this later, but
for now we just avoid it by including the files in our base git repo.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>