Publish blog as Dockerfile
For ease of deploying via Kubernetes, add a Dockerfile so we can publish
a container image containing nginx and the `public` directory.

As a result, we can stop including `public` in our git repo, and instead
only generate it when creating the image.

Currently, building, publishing, and deploying the container image is a
manual process. I'll invest in a CI/CD pipeline as I continue working
with Kubernetes.