Update Dockerfile to manually install Caddy
Previously, we first built a specific
[caddy-docker](https://github.com/abiosoft/caddy-docker) image with our
given plugins via a Make step, and then used said image as the base of
our image. I feel having to do a separate step in Make was a bit of a
smell.

Refactor to perform a multi-stage Docker build, in which we install
Caddy with the necessary plugins in the first stage, and then copy the
binary over to the final stage, along with our static assets and
configuration files. The multi-stage build process has the added
advantage of allowing us to use a distroless build, which decreases
image size and attack surface.

We verify its working via `docker run -p 8080:80
mattjmcnaughton:blog:...`.