Reducing cost of running k8s cluster: part 0
Complete part 0 of the Reducing the Cost of Running a Personal k8s
cluster series. Add outline for the final three parts of the series,
although they are currently marked as drafts.

I split this into multiple components, because it was too much
information for one post (particularly with reducing the cost of the EC2
instances).