Update `mainroad` theme
Update the `mainroad` theme. I don't believe this requires any changes
to my code. Predominantly updating because there's a security
vulnerability in the old version of `es-lint` which `mainroad` was
using. It didn't _really_ matter because I was never installing these
packages.json, but there's no harm in updating.

Also, in general its just good to keep up to date with updates to the
mainroad theme.