chore: upgrade mainroad
Upgrade to the most recent version of mainroad.

The updated version works with `hugo 0.126.1`, which is the version in
nix-pkgs 24.05.

To update, we just delete version in `themes/mainroad` and then run `git
clone` from the mainroad theme again. Then, delete the `.git` file in
the newly cloned directory (because we don't want to mess around w/
submodules).